# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TP2 API de films",
        "description": "API REST pour l’authentification TP2.1",
        "version": "1.0.0"
    },
    "paths": {
        "/api/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Enregistre un nouvel utilisateur",
                "description": "Crée un utilisateur. Ne retourne pas de token. Limite de 5 requêtes/minute.",
                "operationId": "65b7ef1b1e076c033bf91055f53bd9a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "email",
                                    "email_confirmation",
                                    "password",
                                    "password_confirmation",
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "email_confirmation": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur enregistré"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "429": {
                        "description": "Trop de requêtes"
                    }
                }
            }
        },
        "/api/signin": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connecte un utilisateur",
                "description": "Retourne un token et les infos utilisateur. Limite de 5 requêtes/minute.",
                "operationId": "b5cee56757c3c79c5b5f115fdc33f460",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion complete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion complete"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "teyJeEVAaSRVFrSrb"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login ou/et mot de passe invalide"
                    },
                    "429": {
                        "description": "Trop de requêtes"
                    }
                }
            }
        },
        "/api/signout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnecte un utilisateur",
                "description": "Révoque tous les tokens de l’utilisateur. Nécessite un token Bearer. Limite de 5 requêtes/minute.",
                "operationId": "b46f8611fc7d1123f2a3d573cee661c4",
                "responses": {
                    "204": {
                        "description": "Déconnexion réussie (no content)"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "429": {
                        "description": "Trop de requêtes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/films/{id}/critics": {
            "post": {
                "tags": [
                    "Critics"
                ],
                "summary": "Créer une critique pour un film (1 critique max par utilisateur)",
                "operationId": "57de307c4c110984b2fe99d739bde17c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "score"
                                ],
                                "properties": {
                                    "score": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Excellent film !"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Critique créée"
                    },
                    "403": {
                        "description": "Déjà critiqué"
                    },
                    "404": {
                        "description": "Film introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}